requirements.txt
django
djangorestframework
dj-rest-auth
dj-rest-auth[with_social]
python-decouple
djangorestframework-simplejwt
requests
APScheduler